Add location tracking, settings, and UI improvements
Introduces location tracking via foreground service, geofence setup, and location repository. Adds SettingsDataStore, SettingsScreen, and language/dark mode support. Updates permissions in manifest, refactors MainActivity, and improves UI components and navigation. Establishment and review repositories now support rating and review queries. Database version bumped to 7.

+++ app/src/main/java/com/cmu/sweet/ui/state/EstablishmentHomeUiState.kt
@@ -13,5 +13,5 @@ data class EstablishmentHomeUiState(
     val addedBy: String,
     val location: LatLng = LatLng(latitude, longitude),
     val distance: Double? = null,
-    val rating: Float? = null
+    val rating: Double? = null
 )
